add solutions and touch up the comments
