+++ src/test/scala/com/reagroup/exercises/circe/c/CirceEncodingExercisesSpec.scala
@@ -1,0 +1,33 @@
+package com.reagroup.exercises.circe.c
+
+import CirceEncodingExercises._
+import io.circe._
+import io.circe.literal._
+import io.circe.syntax._
+import org.specs2.mutable.Specification
+
+class CirceEncodingExercisesSpec extends Specification {
+  "encodePerson" should {
+
+    "convert Person to Json" in {
+      val person = Person("scala", 20)
+      val actual = encodePerson(person)
+      val expected = Json.obj("name" -> "scala".asJson, "age" -> 20.asJson)
+
+      actual must_=== expected
+    }
+
+  }
+
+  "encodePersonSemiAuto" should {
+
+    "convert Person to Json" in {
+      val person = Person("scala", 20)
+      val actual = encodePersonSemiAuto(person)
+      val expected = Json.obj("name" -> "scala".asJson, "age" -> 20.asJson)
+
+      actual must_=== expected
+    }
+
+  }
+}
